+++ src/components/RemoveListButton.tsx
@@ -4,35 +4,25 @@ import styled from 'styled-components/native';
 import { TouchableOpacity,Text } from 'react-native';
 
 interface ButtonProps {
-    onPress: () => void;
+    onPress: (index: string) => void;
 }
 
 const CreateListButton: FC<ButtonProps> = props => {
     return (
-        <ButtonContainer onPress={props.onPress}>
-           <ButtonText style={{fontWeight: 'bold'}}>+</ButtonText>
-        </ButtonContainer>
+        <ButtonContainer onPress={props.onPress}/>
     )
 }
 export default CreateListButton;
 
 const ButtonContainer = styled.TouchableOpacity`
-    position: absolute;
-    left: 0;
-    top: 0;
-    backgroundColor: #00dc00;
-    width: 90;
-    height: 90;
-    marginTop: 10px;
-    marginRight:5px;
+    alignSelf: flex-end;
+    backgroundColor: red;
+    padding:5px;
+    width: 20;
+    height: 50;
     borderRadius: 50;
     align-items: center;
     justifyContent: center;
     marginLeft: 10px;
     marginBottom: 10px;
 `
-const ButtonText = styled.Text`
-    color: #fff;
-    fontSize: 30;
-`
-