+++ src/entities/todoSingleElement.ts
@@ -1,4 +1,5 @@
 export interface ISingleElementList{
    title:string;
     description: string;
+    index: string;
 }